requirements.txt
jpype1
requests
javaobj-py3
flask

